% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genepairFunctions.R
\name{addCor}
\alias{addCor}
\title{Adds Pearson correlation coefficient for all gene pairs}
\usage{
addCor(gp, expDF, colName = "expCor")
}
\description{
Adds Pearson correlation coefficient for all gene pairs
}
\seealso{
\code{\link{applyToCisPair}}s, \code{\link{applyToCisPairs}}
}
