% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genepairFunctions.R
\name{addGeneAnnotation}
\alias{addGeneAnnotation}
\title{Add annotatino column of each gene to gene pairs}
\usage{
addGeneAnnotation(gp, genesGR, colname)
}
\description{
Add annotatino column of each gene to gene pairs
}
