% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genepairFunctions.R
\name{addPairExp}
\alias{addPairExp}
\title{Adds the expression values for both genes}
\usage{
addPairExp(gp, expDF, expCol, colname = "exp")
}
\arguments{
\item{expDF}{data.frame that can be accesd by rows using the indecies or
names in the first two columns of \code{gp}.}
}
\description{
Adds the expression values for both genes
}
