% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genepairFunctions.R
\name{addSubTADmode}
\alias{addSubTADmode}
\title{Add column to indicate that two query region overlap the same subset of
subject regions. If both query regions do not overlap any subject, FALSE is
returned.}
\usage{
addSubTADmode(gp, tadGR, genesGR, colName = "subTAD")
}
\description{
Add column to indicate that two query region overlap the same subset of
subject regions. If both query regions do not overlap any subject, FALSE is
returned.
}
