% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRangesFunction.R
\name{addWithinSubject}
\alias{addWithinSubject}
\title{Add column to indicate that query lies within at least one subject object.}
\usage{
addWithinSubject(query, subject, colName = "inRegion")
}
\arguments{
\item{query}{A GRanges object}

\item{subject}{A GRanges object}

\item{colName}{A single string that is used as columnname for the new
annotation column in \code{query}.}
}
\value{
A GRanges object like \code{query} with an additioanl column of
  logical values indicating whehter each range is within one range in
  \code{subject} or not.
}
\description{
Add column to indicate that query lies within at least one subject object.
}
