% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomicPairAply.R
\name{applyToCisPairs}
\alias{applyToCisPairs}
\title{Apply a function to a pair of genomic ranges.}
\usage{
applyToCisPairs(gp, rangesGR, datamat, fun = cor)
}
\arguments{
\item{gp}{data.frame of pairs of genomic ranges. First two columns hold
indices of the range in \code{rangesGR}.}

\item{rangesGR}{GenomicRanges object with all ranges used in \code{gp}.}

\item{datamat}{a matrix of with data values associated to each rang in
\code{rangesGR}. Assumes the ranges in rows while \code{row.names(datamat)}
are the \code{names(rangesGR)}.}

\item{fun}{A function that takes two numeric vectors as imput to compute a
summary statsitic. Default is \code{\link{cor()}}.}
}
\value{
A Numeric vector
}
\description{
This function returns a vector with the resulting functin call for each pair..
}
\examples{
rangesGR <- GRanges(
 rep(c("chr1", "chr2"), c(3,2)),
 IRanges(
      c(100, 200, 300, 100, 200),
      c(150, 250, 350, 150, 250)
   ))

datamat <- rbind(
    c(10, 20, 30),
    c(15, 26, 40),
    c(100, 2, 0),
    c(10, 20, 30),
    c(15, 26, 40)
  )
applyToCisPairs(gp, rangesGR, datamat)

}

