% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomicPairAply.R
\name{applyToClosePairs}
\alias{applyToClosePairs}
\title{Apply a function to pairs of close genomic regions.}
\usage{
applyToClosePairs(gp, rangesGR, datamat, fun = cor, maxDist = 10^6)
}
\arguments{
\item{gp}{data.frame of pairs of genomic ranges. First two columns hold
indices of the range in \code{rangesGR}. The rows has to be sorted
according to the first two columns.}

\item{rangesGR}{\code{\link{GenomicRanges}} object with all ranges used in \code{gp}. It
shoul be annoated with \code{\link{seqinfo}} and \code{\link{seqnames}}.}

\item{datamat}{a matrix of with data values associated to each rang in
\code{rangesGR}. Assumes the ranges in rows while \code{row.names(datamat)}
are the \code{names(rangesGR)}.}

\item{fun}{A function that takes two numeric vectors as imput to compute a
summary statsitic. Default is \code{\link{cor()}}.}

\item{maxDist}{maximal distance of pairs in bp as numeric.}
}
\value{
A Numeric vector
}
\description{
This function returns a vector with the resulting functin call for each input
pair.
}
\examples{
rangesGR <- GRanges(
 rep(c("chr1", "chr2"), c(3,2)),
 IRanges(
      c(100, 200, 300, 100, 200),
      c(150, 250, 350, 150, 250)
   ))
gp <- data.frame(
 g1=c(1,4,2,1,4),
 g2=c(2,4,3,3,5)
)

datamat <- rbind(
    c(10, 20, 30),
    c(15, 26, 40),
    c(100, 2, 0),
    c(10, 20, 30),
    c(15, 26, 40)
  )
applyToClosePairs(gp, rangesGR, datamat)

}
