% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genepairFunctions.R
\name{containsGenePairs}
\alias{containsGenePairs}
\title{Test if gene pairs are contained in another set of gene pairs.}
\usage{
containsGenePairs(gp, other)
}
\arguments{
\item{gp}{a data.frame with gene pairs.}

\item{other}{another data.frame with gene pairs.}
}
\value{
a logical vector for each pair in \code{gp} indicating whether it is
  contained in \code{other} or not.
}
\description{
Compaires gene pair by combintion of IDs by using the
\code{\link{getPairIDsorted}} function.
}
