% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genepairFunctions.R
\name{filterForCisPairs}
\alias{filterForCisPairs}
\title{Get subset of gene pairs that are located on the same chromosome.}
\usage{
filterForCisPairs(gp, genesGR)
}
\arguments{
\item{gp}{A data.frame like object holding gene pairs.}

\item{genesGR}{A \code{\link{GRanges}} object.}
}
\value{
A data.frame objec with subset of rows in \code{gp} with pair on the
 same chromosome.
}
\description{
Get subset of gene pairs that are located on the same chromosome.
}
