\name{GenePair-package}
\alias{GenePair-package}
\alias{GenePair}
\docType{package}
\title{
\packageTitle{GenePair}
}
\description{
\packageDescription{GenePair}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{GenePair}
\packageIndices{GenePair}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{GenePair}

Maintainer: \packageMaintainer{GenePair}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
