% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genepair.R
\docType{package}
\name{genepair}
\alias{genepair}
\alias{genepair-package}
\title{genepair: Functions to Analyse Pairs of Genomic Elements like Genes.}
\description{
The genepair package is a collection of functions that work on pairs of
genes. Thse gene pairs consists usually of a data.frame that hold indices of
genes in an associate GRanges object (or strings correponding to names in a
GRanges object).
}
