% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genepairFunctions.R
\name{getAllCisPairs}
\alias{getAllCisPairs}
\title{Get all possible gene pairs within a given distance range as data.frame}
\usage{
getAllCisPairs(genesGR, maxDist = 10^6, minDist = 0)
}
\arguments{
\item{genesGR}{A \code{\link{GRanges}} object.}

\item{maxDist}{single numeric value for maximal allowed distance between
start position.}

\item{minDist}{single numeric value for minimal allowed distance between
start postions.}
}
\value{
A data.frame with
}
\description{
Get all possible gene pairs within a given distance range as data.frame
}
