% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genepairFunctions.R
\name{getCor}
\alias{getCor}
\title{Returns the Pearson correlation coefficient for expression of two input
genes.}
\usage{
getCor(gp, expDF)
}
\description{
If more than one gene pair is provied in \code{gp} a pairwise matirx of
correlation coeficients is returned.
}
