% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genepairFunctions.R
\name{getPairAsGR}
\alias{getPairAsGR}
\title{Make GRange object of range between start position of paired genes.}
\usage{
getPairAsGR(gp, genesGR)
}
\arguments{
\item{gp}{A data.frame like object holding gene pairs.}

\item{genesGR}{A \code{\link{GRanges}} object.}
}
\value{
A \code{\link{GRanges}} object with the range of start positions of
  paired ranges.
}
\description{
It assumes that genes in pair are on the same chromosome.
}
