% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genepairFunctions.R
\name{getPairAsGRL}
\alias{getPairAsGRL}
\title{Make GRangesList object of range between start position of paired genes.}
\usage{
getPairAsGRL(gp, genesGR)
}
\arguments{
\item{gp}{A data.frame like object holding gene pairs.}

\item{genesGR}{A \code{\link{GRanges}} object.}
}
\value{
A \code{\link{GRangesList}} object with the range of start positions of
  paired ranges. GRanges between different chromosomes are empty.
}
\description{
Does NOT assume that genes in pair are on the same chromosome.
}
\seealso{
\code{\link{getPairsAsGRL}}
}
