% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genepairFunctions.R
\name{getPairID}
\alias{getPairID}
\title{Creates an ID for each gene pair by concatenating both gene names.}
\usage{
getPairID(gp)
}
\description{
Creates an ID for each gene pair by concatenating both gene names.
}
