% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genepairFunctions.R
\name{getPairIDsorted}
\alias{getPairIDsorted}
\title{Returns a unique gene ID for each pair that is independent of pair
order (e.g. sorted)}
\usage{
getPairIDsorted(gP)
}
\description{
Returns a unique gene ID for each pair that is independent of pair
order (e.g. sorted)
}
