% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genepairFunctions.R
\name{interChromPairMatrix}
\alias{interChromPairMatrix}
\title{Inter-chromosomal gene pairs counts matrix}
\usage{
interChromPairMatrix(gp, genesGR, symmetric = FALSE)
}
\description{
Inter-chromosomal gene pairs counts matrix
}
