% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genepairFunctions.R
\name{nonOverlappingGenePairs}
\alias{nonOverlappingGenePairs}
\title{Test if gene pairs are non-overlapping. That is that two paird genes do not
overlap each other. in the genome (on the same strand).}
\usage{
nonOverlappingGenePairs(gp, genesGR, useIDs = FALSE)
}
\description{
Test if gene pairs are non-overlapping. That is that two paird genes do not
overlap each other. in the genome (on the same strand).
}
