% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genepairFunctions.R
\name{pairNotNA}
\alias{pairNotNA}
\title{Checks that none of the genes in pairs is NA.}
\usage{
pairNotNA(gp)
}
\value{
a logical vector that has TRUE for each gene pair that has no NA and
  FALSE otherwise.
}
\description{
Checks that none of the genes in pairs is NA.
}
