% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genepairFunctions.R
\name{percentIncluded}
\alias{percentIncluded}
\title{Returns the percentage of gene pairs that are contained in another set of
gene pairs.}
\usage{
percentIncluded(gPa, gPb)
}
\description{
Returns the percentage of gene pairs that are contained in another set of
gene pairs.
}
