% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomization.R
\name{samplePairsByWeight}
\alias{samplePairsByWeight}
\title{Sample pairs from the input gene pairs according to given sampling weights.}
\usage{
samplePairsByWeight(gp, weights, n)
}
\arguments{
\item{gp}{a data.frame with gene pairs. Holds names or indices of ranges in
an associated GRangs object in the first two columns.}

\item{weights}{numeric vector of same length as rows in \code{gp} with
sampling weights.}

\item{n}{the number of pairs to sample.}
}
\value{
a data.frame object with sampled gene pairs.
}
\description{
Sample pairs from the input gene pairs according to given sampling weights.
}
