% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genepairFunctions.R
\name{uniquePair}
\alias{uniquePair}
\title{Remove duplicated entries of gene pairs of the form A-B and B-A.}
\usage{
uniquePair(gp)
}
\description{
Remove duplicated entries of gene pairs of the form A-B and B-A.
}
