% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genepairFunctions.R
\name{uniquePairPerGeneBySim}
\alias{uniquePairPerGeneBySim}
\title{Get only one pair per unique gene by choosing the gene pair with highest
sequence similarity first.}
\usage{
uniquePairPerGeneBySim(gp, similarity)
}
\arguments{
\item{gp}{data.frame like object with gene pairs}

\item{similarty}{numeric vector of the same lenght as rows in \code{gp}}
}
\value{
a data frame that is filtered in a way that each gene occures only in
on pair. The pairs with samller similarity values are keeped preferentially.
}
\description{
Get only one pair per unique gene by choosing the gene pair with highest
sequence similarity first.
}
