% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomization.R
\name{weightsByBin}
\alias{weightsByBin}
\title{Compute sampling weights.}
\usage{
weightsByBin(obs, population, breaks = 50)
}
\arguments{
\item{obs}{vector of observed values according to which the sampling should
be done (e.g distances observed for paralog gene pairs).}

\item{population}{vector of all values in the total population from which one
wants to sample (e.g distances of all gene pairs)}

\item{breaks}{breaks used for sampling resolution (see \code{breaks} argument
in \code{\link[graphics]{hist}} function).}
}
\value{
numeric vector with sampling weights to sample from \code{samp} with
  probabilities observed in \code{obs}.
}
\description{
Compute sampling weights to sample from a population with osberved
probabilities by binning.
}
