% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genepairFunctions.R
\name{addPairDist}
\alias{addPairDist}
\title{Add linear distance between gene paires or NA if on differnet chromosomes.}
\usage{
addPairDist(gp, genesGR, colname = "dist", ignore.strand = FALSE)
}
\description{
Distance is caluclated based on start positon of genes in basepair and can be
negative, if second gene has a lower start postion than the first gene in
pairs.
}
