% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genepairFunctions.R
\name{containsGenePairs}
\alias{containsGenePairs}
\title{Test if gene pairs are contained in another set of gene pairs.}
\usage{
containsGenePairs(gp, negPairs, gPidx = FALSE, nPidx = FALSE, gr = NULL)
}
\value{
a logical vector
}
\description{
Test if gene pairs are contained in another set of gene pairs.
}
