% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomization.R
\name{getRandomPairs}
\alias{getRandomPairs}
\title{Sample randomly pairs from set of genes with equal probabilities}
\usage{
getRandomPairs(genesGR, n)
}
\arguments{
\item{genesGR}{a \code{\link[GenomicRanges]{GRanges}} with genes.}

\item{n}{number of pairs to sample.}
}
\description{
Sample randomly pairs from set of genes with equal probabilities
}
